Refactor fdv.lol app: improved cross-platform CSS and updated spark metric to show red/green based on coin performance
