+++ index.html
@@ -1,62 +1,62 @@
-<!doctype html>
+<!DOCTYPE html>
 <html lang="en">
-<head>
-<meta charset="utf-8" />
-<meta name="viewport" content="width=device-width,initial-scale=1" />
-<title>FDV Memecoin Radar - UltraLite</title>
-<meta name="description" content="Blazing-fast client-side SOL memecoin radar with explainable BUY/WATCH/AVOID." />
-<meta name="author" content="builders-toronto" />
-<meta name="keywords" content="solana, memecoin, bonk, wif, dog, inu, pepe, cat, meme, ponk, samo, purry, purr, kitty, meow, woof, frog, snek, toad, bob, dino, monke, monkey, ape, corgi, floki, elon" />
-<meta property="og:title" content="FDV Memecoin Radar - UltraLite" />
-<meta property="og:description" content="Blazing-fast client-side SOL memecoin radar with explainable BUY/WATCH/AVOID." />
-<meta property="og:image" content="assets/images/fdv.lol.png" />
-<meta property="og:image:width" content="1200" />
-<meta property="og:image:height" content="630" />
-<meta property="og:type" content="website" />
+  <head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width,initial-scale=1" />
+    <title>FDV Memecoin Radar - UltraLite</title>
+    <meta name="description" content="Blazing-fast client-side SOL memecoin radar with explainable BUY/WATCH/AVOID." />
+    <meta name="author" content="builders-toronto" />
+    <meta name="keywords" content="solana, memecoin, bonk, wif, dog, inu, pepe, cat, meme, ponk, samo, purry, purr, kitty, meow, woof, frog, snek, toad, bob, dino, monke, monkey, ape, corgi, floki, elon" />
+    <meta property="og:title" content="FDV Memecoin Radar - UltraLite" />
+    <meta property="og:description" content="Blazing-fast client-side SOL memecoin radar with explainable BUY/WATCH/AVOID." />
+    <meta property="og:image" content="assets/images/fdv.lol.png" />
+    <meta property="og:image:width" content="1200" />
+    <meta property="og:image:height" content="630" />
+    <meta property="og:type" content="website" />
 
-<link rel="icon" type="image/png" href="assets/images/icons/fdv.lol.ico" />
-<link rel="preconnect" href="https://fonts.googleapis.com">
-<link href="https://fonts.googleapis.com/css2?family=Inter:wght@400;600&display=swap" rel="stylesheet">
-<link rel="stylesheet" href="assets/styles/global.css" />
-</head>
-<body>
-<div id="loader" class="loader-overlay" aria-live="polite" aria-busy="true" hidden>
-  <div class="loader-wrap">
-    <span class="sr-only">Loading…</span>
-    <svg class="spinner" viewBox="0 0 50 50" role="img" aria-label="Loading">
-      <circle class="ring" cx="25" cy="25" r="22" />
-      <circle class="arc"  cx="25" cy="25" r="22" />
-    </svg>
+    <link rel="icon" type="image/png" href="assets/images/icons/fdv.lol.ico" />
+    <link rel="preconnect" href="https://fonts.googleapis.com">
+    <link href="https://fonts.googleapis.com/css2?family=Inter:wght@400;600&display=swap" rel="stylesheet">
+    <link rel="stylesheet" href="src/styles/global.css" />
+  </head>
+  <body>
+  <div id="loader" class="loader-overlay" aria-live="polite" aria-busy="true" hidden>
+    <div class="loader-wrap">
+      <span class="sr-only">Loading…</span>
+      <svg class="spinner" viewBox="0 0 50 50" role="img" aria-label="Loading">
+        <circle class="ring" cx="25" cy="25" r="22" />
+        <circle class="arc"  cx="25" cy="25" r="22" />
+      </svg>
+    </div>
   </div>
-</div>
-<header class="header">
-  <div class="container">
-    <h1>⚡ FDV lol</h1>
-    <div id="meta" class="sub">Loading…</div>
-    <div class="controls">
-      <select id="sort">
-        <option value="score">Sort: Score</option>
-        <option value="volume">Sort: 24h Volume</option>
-        <option value="liquidity">Sort: Liquidity</option>
-        <option value="change24">Sort: 24h Change</option>
-      </select>
-      <div class="innerControls">
-        <button id="refresh" class="btn">Refresh</button>
-        <div class="relaxControl">
-            relax filter
-            <input type="checkbox" id="relax" disabled/>
+  <header class="header">
+    <div class="container">
+      <h1>⚡ FDV lol</h1>
+      <div id="meta" class="sub">Loading…</div>
+      <div class="controls">
+        <select id="sort">
+          <option value="score">Sort: Score</option>
+          <option value="volume">Sort: 24h Volume</option>
+          <option value="liquidity">Sort: Liquidity</option>
+          <option value="change24">Sort: 24h Change</option>
+        </select>
+        <div class="innerControls">
+          <button id="refresh" class="btn">Refresh</button>
+          <div class="relaxControl">
+              relax filter
+              <input type="checkbox" id="relax" disabled/>
+          </div>
         </div>
       </div>
     </div>
-  </div>
-</header>
+  </header>
 
-<main class="container grid">
-  <div id="cards" class="cards"></div>
-  <footer>
-    <p>Cached in your browser for 90s. Not financial advice.</p><p><a href="https://github.com/builders-toronto/fdv.lol" target="_blank" rel="noopener">Source code</a></p><p><span class="joinCta">Join us</span>: <a href="https://t.me/fdvlol" target="_blank" rel="noopener">Telegram</a> or <a href="https://x.com/lfbuilders" target="_blank" rel="noopener">X</a><br/></p>
-  </footer>
-</main>
-<script src="assets/scripts/app.js"></script>
-</body>
+  <main class="container grid">
+    <div id="cards" class="cards"></div>
+    <footer>
+      <p>Cached in your browser for 90s. Not financial advice.</p><p><a href="https://github.com/builders-toronto/fdv.lol" target="_blank" rel="noopener">Source code</a></p><p><span class="joinCta">Join us</span>: <a href="https://t.me/fdvlol" target="_blank" rel="noopener">Telegram</a> or <a href="https://x.com/lfbuilders" target="_blank" rel="noopener">X</a><br/></p>
+    </footer>
+  </main>
+  <script type="module" src="app.js"></script>
+  </body>
 </html>
