+++ src/styles/global.css
@@ -219,7 +219,7 @@ input[type="checkbox"]:disabled{
 .pill.down{color:#ff98ad; border-color:rgba(255,77,109,.35)}
 .tagline{color:var(--muted); font-size:.88rem; margin:6px 0 10px}
 
-.actions{display:flex; gap:10px; align-items:center; flex-wrap:wrap;justify-content: flex-end;}
+.actions{display:flex; gap:10px; align-items:center; flex-wrap:wrap;justify-content: space-between; margin-top: 15px;}
 
 .swapCoin{
   border: 2px solid cyan !important;
@@ -271,14 +271,11 @@ input[type="checkbox"]:disabled{
 }
 
 .card .micro {
-  display: flex;
-  flex-direction: column;
-  align-items: center !important;
-  position: absolute; 
-  left: 10px; 
-  bottom: 8px;
-  gap: 8px; 
-  align-items: end;
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    justify-content: space-between;
+    margin-top: 26px;
 }
 .micro .spark {
   width: 120px; height: 32px; display: block; opacity: .9;
@@ -303,7 +300,7 @@ input[type="checkbox"]:disabled{
   position:absolute; left:0; right:0; bottom:-14px; font-size:10px; text-align:center; color:#9acfee90;
 }
 
-.card .actions:last-child { margin-bottom: 50px; }
+/* .card .actions:last-child { margin-bottom: 50px; } */
 
 .micro .pctrow {
   display: flex; gap: 6px; align-items: center;
@@ -436,7 +433,6 @@ input[type="checkbox"]:disabled{
 @media (prefers-reduced-motion: reduce){
   .adcard::after, .adlogo, .adtag::after{ animation:none }
 }
-
 
 footer {
     display: flex;
